Factor out LogSliderWrapper
